requirements.txt
numpy==1.16.5
matplotlib==3.3.4
torchvision==0.5.0
torch==1.4.0
Pillow==8.1.2
transformers==4.4.2
